% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWEV.R
\name{dWEV}
\alias{dWEV}
\alias{recalc_t0}
\alias{WEVmodel}
\alias{dynWEV}
\title{Dynamical weighted evidence and visibility model (dynWEV)}
\usage{
recalc_t0(t0, st0)

dWEV(rt, th1, th2, response = "upper", a, v, t0, z = 0.5, d = 0,
  sz = 0, sv = 0, st0 = 0, tau = 1, w = 0.5, muvis = NULL,
  sigvis = 0, svis = 1, s = 1, simult_conf = FALSE,
  precision = 1e-05, z_absolute = FALSE, stop_on_error = TRUE)
}
\arguments{
\item{t0}{non-decision time or response time constant (in seconds). Lower bound for the duration of all non-decisional processes (encoding and response execution). Typical range: 0.1 < \code{t0} < 0.5}

\item{st0}{inter-trial-variability of non-decisional components. Range of a uniform distribution with mean \code{t0 + st0/2} describing the distribution of actual \code{t0} values across trials. Accounts for response times below \code{t0}. Reduces skew of predicted RT distributions. Values different from 0 can slow computation considerably. Typical range: 0 < \code{st0} < 0.2. Default is 0.}

\item{rt}{a vector of RTs. Or for convenience also a \code{data.frame} with columns \code{rt} and \code{response} (such as returned from \code{rdiffusion} or \strong{rLBA}). See examples.}

\item{th1}{together with th2: scalars or numerical vectors giving the lower and upper bound of the interval, in which the accumulator should end at the time of the confidence judgement (i.e. at time \code{rt}+\code{tau}). Only values with \code{th2}>=\code{th1} are accepted.}

\item{th2}{(see th1)}

\item{response}{character vector. Which response boundary should be tested? Possible values are \code{c("upper", "lower")}, possibly abbreviated and \code{"upper"} being the default. Alternatively, a numeric vector with values 1=lower and 2=upper. For convenience, \code{response} is converted via \code{as.numeric} also allowing factors (see examples). Ignored if the first argument is a \code{data.frame}.}

\item{a}{threshold separation. Amount of information that is considered for a decision. Large values indicate a conservative decisional style. Typical range: 0.5 < \code{a} < 2}

\item{v}{drift rate. Average slope of the information accumulation process. The drift gives information about the speed and direction of the accumulation of information. Large (absolute) values of drift indicate a good performance. If received information supports the response linked to the upper threshold the sign will be positive and vice versa. Typical range: -5 < \code{v} < 5}

\item{z}{starting point. Indicator of an a priori bias in decision making. When the relative starting point \code{z} deviates from \code{0.5*a}, the amount of information necessary for a decision differs between response alternatives. Default is \code{0.5*a} (i.e., no bias).}

\item{d}{differences in speed of response execution (in seconds). Positive values indicate that response execution is faster for responses linked to the upper threshold than for responses linked to the lower threshold. Typical range: -0.1 < \code{d} < 0.1. Default is 0.}

\item{sz}{inter-trial-variability of starting point. Range of a uniform distribution with mean \code{z} describing the distribution of actual starting points from specific trials. Values different from 0 can predict fast errors (but can slow computation considerably). Typical range: 0 < \code{sz} < 0.2. Default is 0. (Given in relative range i.e. bounded by 2*min(z, 1-z))}

\item{sv}{inter-trial-variability of drift rate. Standard deviation of a normal distribution with mean \code{v} describing the distribution of actual drift rates from specific trials. Values different from 0 can predict slow errors. Typical range: 0 < \code{sv} < 2. Default is 0.}

\item{tau}{postdecisional accumulation time; the length of the time period after the decision was made until the confidence judgement is made. Range: \code{tau}>0. Default: \code{tau}=1.}

\item{w}{weight put on the final state of the decision accumulator. 1-w is the weight on the visibility accumulator. Range: 0<\code{w}<1. Default: \code{w}=0.5.}

\item{muvis}{mean drift of visibility accumulation process. Default: \code{muvis}=1.}

\item{sigvis}{the variability in drift rate of the visibility accumulator (which varies independently from the drift rate in decision process). Range: \code{sigvis}>=0. Default: \code{sigvis}=0.}

\item{svis}{diffusion constant of visibility accumulation process. Range: \code{svis}>0. Default: \code{svis}=1.}

\item{s}{diffusion constant of decision process; standard deviation of the random noise of the diffusion process (i.e., within-trial variability), scales other parameters (see Note). Needs to be fixed to a constant in most applications. Default is 1. Note that the default used by Ratcliff and in other applications is often 0.1.}

\item{simult_conf}{logical. Whether in the experiment confidence was reported simultaneously
with the decision, as then decision and confidence judgment are assumed to have happened
subsequent before response and computations are different, when there is an observable
interjudgment time (then simult_conf should be FALSE).}

\item{precision}{\code{numerical} scalar value. Precision of calculation. #' @param precision \code{numerical} scalar value. Precision of calculation. Corresponds to the stepsize of integration w.r.t. z and t0. Default is 1e-5.}

\item{z_absolute}{logical. Determines whether z is treated as absolute start point (TRUE) or relative (FALSE; default) to a.}

\item{stop_on_error}{Should the diffusion functions return 0 if the parameters values are outside the allowed range (= \code{FALSE}) or produce an error in this case (= \code{TRUE}).}
}
\value{
\code{d2DSD(WEV(d/mu/absmu))} gives the density/likelihood/probability of the diffusion process producing a decision of \code{response} at time \code{rt} and a confidence judgment corresponding to the interval \[ \code{th1}, \code{th2}\]. The value will be a numeric vector of the same length as \code{rt}.

The distribution parameters (as well as \code{response}, \code{tau}, \code{th1} and \code{th2}, \code{w} and \code{sig}) are recycled to the length of the result. In other words, the functions are completely vectorized for all parameters and even the response boundary.
}
\description{
Likelihood function for the dynWEV model (Hellmann & Rausch).
It includes following parameters from the drift diffusion model:
\code{a} (threshold separation),
\code{z} (starting point; relative),
\code{v} (drift rate),
\code{t0} (non-decision time/response time constant),
\code{d} (differences in speed of response execution),
\code{sv} (inter-trial-variability of drift),
\code{st0} (inter-trial-variability of non-decisional components),
\code{sz} (inter-trial-variability of relative starting point) and
\code{s} (diffusion constant).
For the computation of confidence following parameters were added:
\code{tau} (postdecisional accumulation time),
\code{w} (weight on the decision evidence (weight on visibility is (1-w))),
\code{muvis} (mean drift rate of visibility process),
\code{svis} (diffusion constant of visibility process),
\code{sigvis} (variability in drift rate of visibility accumulator), and
\code{th1} and \code{th2} (lower and upper thresholds for confidence interval).
 \strong{Note that the parameterization or defaults of non-decision time variability
 \code{st0} and diffusion constant \code{s} differ from what is often found in the literature.}
 The code is basically an extension of the \code{\link[rtdists:Diffusion]{ddiffusion}}
 function for the Ratcliff diffusion model.
}
\details{
The function is a extension of the \code{ddiffusion} function from the package \code{rtdists} for the Ratcliff diffusion model. The Ratcliff diffusion model (Ratcliff, 1978) is a mathematical model for two-choice discrimination tasks. It is based on the assumption that information is accumulated continuously until one of two decision thresholds is hit. For introductions see Ratcliff and McKoon (2008), Voss, Rothermund, and Voss (2004), Voss, Nagler, and Lerche (2013), or Wagenmakers (2009).

The dynWEV model is an extension of the drift diffusion model that incorporates the idea of an additional process accumulating information about the general visibility (also decision irrelevant features) of the stimulus as well as a postdecisional accumulation period (\strong{Rausch, Hellmann & Zehetleitner (2018)}).

All functions are fully vectorized across all parameters as well as the response to match the length or \code{rt} (i.e., the output is always of length equal to \code{rt}). This allows for trialwise parameters for each model parameter.

For convenience, the function allows that the first argument is a \code{data.frame} containing the information of the first and second argument in two columns (i.e., \code{rt} and \code{response}). Other columns (as well as passing \code{response} separately argument) will be ignored.
}
\note{
The parameterization of the non-decisional components, \code{t0} and \code{st0}, differs from the parameterization used by, for example, Andreas Voss or Roger Ratcliff. In the present case \code{t0} is the lower bound of the uniform distribution of length \code{st0}, but \emph{not} its midpoint. The parameterization employed here is in line with the parametrization for the \strong{LBA} code (where \code{t0} is also the lower bound).

The default diffusion constant \code{s} is 1 and not 0.1 as in most applications of Roger Ratcliff and others. Usually \code{s} is not specified as the other parameters:
\code{a}, \code{v}, \code{sv}, \code{muvis}, \code{sigvis}, and \code{svis} respectively, may be scaled to produce the same distributions (as is done in the code).
}
\references{
Ratcliff, R. (1978). A theory of memory retrieval. \emph{Psychological Review}, 85(2), 59-108.

Ratcliff, R. (1978). A theory of memory retrieval. \emph{Psychological Review}, 85(2), 59-108.

Ratcliff, R., & McKoon, G. (2008). The diffusion decision model: Theory and data for two-choice decision tasks. \emph{Neural Computation}, 20(4), 873-922.

Rausch, M., Hellmann, S., & Zehetleitner, M. (2018). Confidence in masked orientation judgments is informed by both evidence and visibility. \emph{Attention, Perception, & Psychophysics}, 80(1), 134–154.  doi: 10.3758/s13414-017-1431-5

Voss, A., Rothermund, K., & Voss, J. (2004). Interpreting the parameters of the diffusion model: An empirical validation. \emph{Memory & Cognition}. Vol 32(7), 32, 1206-1220.

Voss, A., Nagler, M., & Lerche, V. (2013). Diffusion Models in Experimental Psychology: A Practical Introduction. \emph{Experimental Psychology}, 60(6), 385-402. doi:10.1027/1618-3169/a000218

Wagenmakers, E.-J., van der Maas, H. L. J., & Grasman, R. P. P. P. (2007). An EZ-diffusion model for response time and accuracy. \emph{Psychonomic Bulletin & Review}, 14(1), 3-22.

Wagenmakers, E.-J. (2009). Methodological and empirical developments for the Ratcliff diffusion model of response times and accuracy. \emph{European Journal of Cognitive Psychology}, 21(5), 641-671.
}
\author{
For the original rtdists package: Underlying C code by Jochen Voss and Andreas Voss. Porting and R wrapping by Matthew Gretton, Andrew Heathcote, Scott Brown, and Henrik Singmann. \code{qdiffusion} by Henrik Singmann. For the d2DSD function the C code was extended by Sebastian Hellmann.
}
