% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWEV.R
\name{dynWEV}
\alias{dynWEV}
\alias{dWEV}
\alias{WEVmodel}
\alias{ddynWEV}
\alias{rWEV}
\title{Dynamical weighted evidence and visibility model (dynWEV)}
\usage{
dWEV(rt, th1, th2, response = "upper", a, v, t0 = 0, z = 0.5, d = 0,
  sz = 0, sv = 0, st0 = 0, tau = 1, w = 0.5, muvis = NULL,
  sigvis = 0, svis = 1, s = 1, simult_conf = FALSE,
  precision = 1e-05, z_absolute = FALSE, stop_on_error = TRUE)

rWEV(n, a, v, t0 = 0, z = 0.5, d = 0, sz = 0, sv = 0, st0 = 0,
  tau = 1, w = 0.5, muvis = NULL, sigvis = 0, svis = 1, s = 1,
  delta = 0.01, maxrt = 15, simult_conf = FALSE, z_absolute = FALSE,
  stop_on_error = TRUE)
}
\arguments{
\item{rt}{a vector of RTs. Or for convenience also a \code{data.frame} with columns \code{rt}
and \code{response}.}

\item{th1}{together with th2: scalars or numerical vectors giving the lower and upper bound of the
interval of the confidence measure (see Details). Only values with \code{th2}>=\code{th1} are accepted.}

\item{th2}{(see th1)}

\item{response}{character vector, indicating the decision, i.e. which boundary was
met first. Possible values are \code{c("upper", "lower")} (possibly abbreviated) and
\code{"upper"} being the default. Alternatively, a numeric vector with values 1=lower
and 2=upper. For convenience, \code{response} is converted via \code{as.numeric} also
allowing factors. Ignored if the first argument is a \code{data.frame}.}

\item{a}{threshold separation. Amount of information that is considered for a decision. Large values
indicate a conservative decisional style. Typical range: 0.5 < \code{a} < 2}

\item{v}{drift rate of decision process. Average slope of the information accumulation
process. The drift gives information about the speed and direction of the accumulation of information.
Large (absolute) values of drift indicate a good performance. If received information supports the
response linked to the upper threshold the sign will be positive and vice versa.
Typical range: -5 < \code{v} < 5}

\item{t0}{non-decision time or response time constant (in seconds). Lower bound for the duration of all
non-decisional processes (encoding and response execution).
Typical range: 0.1 < \code{t0} < 0.5. Default is 0.}

\item{z}{(by default relative) starting point of decision process. Indicator of an a priori bias in decision
making. When the relative starting point \code{z} deviates from \code{0.5}, the amount
of information necessary for a decision differs between response alternatives. Default
is \code{0.5} (i.e., no bias).}

\item{d}{differences in speed of response execution (in seconds). Positive values indicate that
response execution is faster for responses linked to the upper threshold than for responses linked to
the lower threshold. Typical range: -0.1 < \code{d} < 0.1. Default is 0.}

\item{sz}{inter-trial-variability of starting point. Range of a uniform distribution with mean
\code{z} describing the distribution of actual starting points from specific trials. Values different
from 0 can predict fast errors (but can slow computation considerably). Typical range:
0 < \code{sz} < 0.2. Default is 0. (Given in relative range i.e. bounded by 2*min(z, 1-z))}

\item{sv}{inter-trial-variability of drift rate of decision process. Standard deviation of a normal
distribution with mean \code{v} describing the distribution of actual drift rates from specific trials.
Values different from 0 can predict slow errors. Typical range: 0 < \code{sv} < 2. Default is 0.}

\item{st0}{inter-trial-variability of non-decisional components. Range of a uniform distribution with
mean \code{t0 + st0/2} describing the distribution of actual \code{t0} values across trials.
Accounts for response times below \code{t0}. Reduces skew of predicted RT distributions.
Values different from 0 can slow computation considerably. Typical range: 0 < \code{st0} < 0.2.
Default is 0.}

\item{tau}{post-decisional accumulation time; the length of the time period after the decision was
made until the confidence judgment is made. Range: \code{tau}>0. Default: \code{tau}=1.}

\item{w}{weight put on the final state of the decision accumulator for confidence computation.
1-w is the weight on the visibility accumulator. Range: 0<\code{w}<1. Default: \code{w}=0.5.}

\item{muvis}{mean drift of visibility process. If \code{NULL} (default), \code{muvis} will be set to the
absolute value of \code{v}.}

\item{sigvis}{the variability in drift rate of the visibility process (which varies independently
from the drift rate in decision process). Range: \code{sigvis}>=0. Default: \code{sigvis}=0.}

\item{svis}{diffusion constant of visibility process. Range: \code{svis}>0. Default: \code{svis}=1.}

\item{s}{diffusion constant of decision process; standard deviation of the random noise of the
diffusion process (i.e., within-trial variability), scales other parameters (see Note). Needs to
be fixed to a constant in most applications. Default is 1. Note that the default used by Ratcliff
and in other applications is often 0.1.}

\item{simult_conf}{logical. Whether in the experiment confidence was reported simultaneously
with the decision. If that is the case decision and confidence judgment are assumed to have happened
subsequent before the response. Therefore \code{tau} is included in the response time. If the decision was
reported before the confidence report, \code{simul_conf} should be \code{FALSE}.}

\item{precision}{numerical scalar value. Precision of calculation. Corresponds to the
step size of integration w.r.t. \code{z} and \code{t0}. Default is 1e-5.}

\item{z_absolute}{logical. Determines whether z is treated as absolute start point (TRUE) or
relative (FALSE; default) to \code{a}.}

\item{stop_on_error}{Should the diffusion functions return 0 if the parameters values are
outside the allowed range (= \code{FALSE}) or produce an error in this case (= \code{TRUE}).}

\item{n}{integer. The number of samples generated.}

\item{delta}{numeric. Discretization step size for simulations in the stochastic process}

\item{maxrt}{numeric. Maximum decision time returned. If the simulation of the stochastic
process exceeds a decision time of \code{maxrt}, the \code{response} will be set to 0 and the \code{maxrt}
will be returned as \code{rt}.}
}
\value{
\code{dWEV} gives the density/likelihood/probability of the diffusion process producing
a decision of \code{response} at time \code{rt} and a confidence judgment corresponding to the
interval [ \code{th1}, \code{th2}]. The value will be a numeric vector of the same length as
\code{rt}.

\code{rWEV} returns a \code{data.frame} with three columns and \code{n} rows. Column names are \code{rt} (response
time), \code{response} (-1 (lower) or 1 (upper), indicating which bound was hit), and \code{conf} (the
value of the confidence measure; not discretized!).

The distribution parameters (as well as \code{response}, \code{tau}, \code{th1} and \code{th2},
\code{w} and \code{sig}) are recycled to the length of the result. In other words, the functions
are completely vectorized for all parameters and even the response boundary.
}
\description{
Likelihood function and random number generator for the dynWEV model
(Hellmann et al., preprint).
It includes following parameters from the drift diffusion model:
\code{a} (threshold separation),
\code{z} (starting point; relative),
\code{v} (drift rate),
\code{t0} (non-decision time/response time constant),
\code{d} (differences in speed of response execution),
\code{sv} (inter-trial-variability of drift),
\code{st0} (inter-trial-variability of non-decisional components),
\code{sz} (inter-trial-variability of relative starting point) and
\code{s} (diffusion constant).
For the computation of confidence following parameters were added:
\code{tau} (post-decisional accumulation time),
\code{w} (weight on the decision evidence (weight on visibility is (1-w))),
\code{muvis} (mean drift rate of visibility process),
\code{svis} (diffusion constant of visibility process),
\code{sigvis} (variability in drift rate of visibility accumulator), and
\code{th1} and \code{th2} (lower and upper thresholds for confidence interval).
\strong{Note that the parametrization or defaults of non-decision time variability
\code{st0} and diffusion constant \code{s} differ from what is often found in the literature.}
}
\details{
The dynamical weighted evidence and visibility model is an extension of the 2DSD model
for decision confidence (see \code{\link{d2DSD}}). It assumes that the decision follows a drift
diffusion model with two additional assumptions to account for confidence. First, there is a
post-decisional period of further evidence accumulation \code{tau}. Second, another accumulation process
accrues information about stimulus reliability (the visibility process) including also evidence
about decision irrelevant features. See Hellmann et al. (preprint) for more information.
The measure for confidence is then a weighted sum of the final state of the decision process X
and the visibility process V, i.e. for a decision time T (which is not the response time),
the confidence measure is
\deqn{conf = wX(T+\tau) + (1-w) V(T+\tau).}

All functions are fully vectorized across all parameters as well as the response to match the
length or \code{rt} (i.e., the output is always of length equal to \code{rt}). This allows for
trial wise parameters for each model parameter.

For convenience, the function allows that the first argument is a \code{data.frame} containing
the information of the first and second argument in two columns (i.e., \code{rt} and \code{response}).
Other columns (as well as passing \code{response} separately argument) will be ignored.
}
\note{
The parameterization of the non-decisional components, \code{t0} and \code{st0},
differs from the parameterization sometimes used in the literature.
In the present case \code{t0} is the lower bound of the uniform distribution of length
\code{st0}, but \emph{not} its midpoint. The parameterization employed here is in line
with the functions in the \code{rtdists} package.

The default diffusion constant \code{s} is 1 and not 0.1 as in most applications of
Roger Ratcliff and others. Usually \code{s} is not specified as the other parameters:
\code{a}, \code{v}, \code{sv}, \code{muvis}, \code{sigvis}, and \code{svis} respectively,
may be scaled to produce the same distributions (as is done in the code).

The function code is basically an extension of the \code{ddiffusion} function from the
package \code{rtdists} for the Ratcliff diffusion model.
}
\references{
Hellmann, S., Zehetleitner, M., & Rausch, M. (preprint). Simultaneous modeling of choice,
confidence and response time in visual perception. https://osf.io/9jfqr/
}
\author{
Sebastian Hellmann
}
