% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaceModels.R
\name{RaceModels}
\alias{RaceModels}
\alias{dIRM}
\alias{dPCRM}
\alias{rRM}
\alias{dRM}
\alias{rIRM}
\alias{rPCRM}
\alias{racemodels}
\title{Independent and partially anti-correlated Race Model for Decision Confidence}
\usage{
dIRM(rt, response = 1, mu1, mu2, a, b, th1, th2, wx = 1, wrt = 0,
  wint = 0, t0 = 0, st0 = 0, s = 1, time_scaled = TRUE,
  step_width = NULL)

dPCRM(rt, response = 1, mu1, mu2, a, b, th1, th2, wx = 1, wrt = 0,
  wint = 0, t0 = 0, st0 = 0, s = 1, time_scaled = TRUE,
  step_width = NULL)

rIRM(n, mu1, mu2, a, b, wx = 1, wrt = 0, wint = 0, t0 = 0, st0 = 0,
  s = 1, time_scaled = TRUE, step_width = NULL, delta = 0.01,
  maxrt = 15)

rPCRM(n, mu1, mu2, a, b, wx = 1, wrt = 0, wint = 0, t0 = 0, st0 = 0,
  s = 1, time_scaled = TRUE, step_width = NULL, delta = 0.01,
  maxrt = 15)
}
\arguments{
\item{rt}{a numeric vector of RTs. For convenience also a \code{data.frame} with
columns \code{rt} and \code{response} is possible.}

\item{response}{numeric vector with values in \code{c(1, 2)}, giving the accumulator that hit its
boundary first.}

\item{mu1}{numeric. Drift rate for the first accumulator}

\item{mu2}{numeric. Drift rate for the second accumulator}

\item{a}{positive numeric. Distance from starting point to boundary of the first accumulator.}

\item{b}{positive numeric. Distance from starting point to boundary of the second accumulator.}

\item{th1}{numeric. Lower bound of interval range for the confidence measure.}

\item{th2}{numeric. Upper bound of interval range for the confidence measure.}

\item{wx}{numeric. Weight on losing accumulator for the computation of the confidence measure.
(Used only if time_scale==TRUE, Default = 1)}

\item{wrt}{numeric. Weight on reaction time for the computation of the confidence measure.
(Used only if time_scale==TRUE, Default wrt=0)}

\item{wint}{numeric. Weight on the interaction of losing accumulator and reaction time for
the computation of the confidence measure. (Used only if time_scale==TRUE,
Default wint=0)}

\item{t0}{numeric. Lower bound of non-decision time component in observable response times.
Range: t0>=0. Default: t0=0.}

\item{st0}{numeric. Range of a uniform distribution for non-decision time. Range: st0>=0.
Default: st0=0.}

\item{s}{numeric. Diffusion constant of the two accumulators.  Usually fixed to 1 for most
purposes as it scales other parameters (see Details). Range: s>0, Default: s=1.}

\item{time_scaled}{logical. Whether the confidence measure should be time-dependent. See Details.}

\item{step_width}{numeric. Step size for the integration in t0 (motor time). Default: 1e-6.}

\item{n}{integer. The number of samples generated.}

\item{delta}{numeric. Discretization step size for simulations in the stochastic process}

\item{maxrt}{numeric. Maximum decision time returned. If the simulation of the stochastic
process exceeds a decision time of \code{maxrt}, the \code{response} will be set to 0 and the \code{maxrt}
will be returned as \code{rt}.}
}
\value{
\code{dIRM} and \code{dPCRM} return the numerical value of the probability density in a numerical vector of the same
length as \code{rt}.

\code{rIRM} and \code{dPCRM} return a \code{data.frame} with four columns and \code{n} rows. Column names are \code{rt} (response
time), \code{response} (1 or 2, indicating which accumulator hit its boundary first),
\code{xl} (the final state of the loosing accumulator), and \code{conf} (the
value of the confidence measure; not discretized!).

The race parameters (as well as \code{response}, \code{th1},
and \code{th2}) are recycled to the length of the result (either \code{rt} or \code{n}).
In other words, the functions are completely vectorized for all parameters
and even the response.
}
\description{
Probability densities and random number generators for response times,
decisions and confidence judgments in the independent Race Model
(\code{dIRM}/\code{rIRM}) or partially (anti-)correlated Race Model (\code{dPCRM}/\code{rPCRM}),
i.e. the probability of a given response (response: winning accumulator
(1 or 2)) at a given time (rt) and the confidence measure in the interval
between th1 and th2 (Hellmann et al., preprint). The definition of the
confidence measure depends on the argument \code{time_scaled} (see Details).
The computations are based on Moreno-Bote (2010).
The parameters for the models are \code{mu1} and \code{mu2} for the drift
rates, \code{a}, \code{b} for the upper thresholds of the two accumulators
and \code{s} for the incremental standard deviation of the processes and
\code{t0} and \code{st0} for the  minimum and range of uniformly distributed
non-decision times (including encoding and motor time).
For the computation of confidence judgments, the parameters \code{th1} and
\code{th2} for the lower and upper bound of the interval for confidence
measure and if \code{time_scaled} is \code{TRUE} the weight parameters \code{wx}, \code{wrt},
\code{wint} for the computation of the confidence measure are required (see Details).
}
\details{
The parameters are formulated, s.t. both accumulators start at 0 and trigger a decision at their
positive boundary \code{a} and \code{b} respectively. That means, both parameters have to be positive.
Internally the computations adapt the parametrization of Moreno-Bote (2010).

\code{time_scaled} determines whether the confidence measure is computed in accordance to the
Balance of Evidence hypothesis (if \code{time_scaled=FALSE}), i.e. if \code{response} is 1 at time T and
\eqn{X_2}{X2} is the second accumulator, then
\deqn{conf = b - X_2(T)}{conf = b - X2(T)}.
Otherwise, if \code{time_scaled=TRUE} (default), confidence is computed as linear combination of
Balance of Evidence, decision time, and an interaction term, i.e.
\deqn{conf = wx (b-X_2 (T)) + wrt\frac{1}{\sqrt{T}} + wint\frac{b-X_2(T)}{\sqrt{T}}.}{conf = wx (b-X2(T)) + wrt/\sqrt{T} + wint (b-X2(T))\sqrt{T}.}
Usually the weights (\code{wx}, \code{wrt}, \code{wint}) should sum to 1, as the confidence thresholds
(\code{th1} and \code{th2}) may be scaled according to their sum. If this is not the case, they will be scaled
accordingly internally! Usually, this formula results in lower confidence when the reaction time is
longer but the state of the second accumulator is held constant. It is based on the optimal decision
confidence in Moreno-Bote (2010).

For convenience, the likelihood function allows that the first argument is a \code{data.frame} containing the
information of the first and second argument in the columns
(i.e., \code{rt} and \code{response}). Other columns (as well as passing
\code{response} separately as argument) will be ignored.

The simulations are done by simulating normal variables in discretized steps until
one process reaches the boundary. If no boundary is met within the maximum time, response is
set to 0.
}
\note{
Similarly to the drift diffusion models (like \code{ddiffusion} and
\code{\link{dWEV}}), s is a scaling factor (scales: \code{mu1},\code{mu2}, \code{a},\code{b},
\code{th1},\code{th2},and \code{wrt}) and is usually fixed to 1.
}
\references{
Hellmann, S., Zehetleitner, M., & Rausch, M. (preprint). Simultaneous modeling of choice,
confidence and response time in visual perception. https://osf.io/9jfqr/

Moreno-Bote, R. (2010). Decision confidence and uncertainty in diffusion models with partially
correlated neuronal integrators. Neural Computation, 22(7), 1786–1811. https://doi.org/10.1162/neco.2010.12-08-930
}
\author{
Sebastian Hellmann
}
