% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRM.R
\name{rRM}
\alias{rRM}
\alias{simulateRM}
\alias{simulateIRM}
\alias{simulatePCRM}
\alias{rRM_CPP}
\alias{rRM_Kiani}
\title{Simulation of confidence ratings and RTs in race confidence models}
\usage{
rRM(paramDf, n = 10000, model = "IRM", time_scaled = FALSE,
  gamma = FALSE, agg_simus = FALSE, stimulus = c(1, 2), delta = 0.01,
  maxrt = 15, seed = NULL)

rRM_CPP(n = 10000, paramDf, model = "IRM", delta = 0.01, maxrt = 15)

rRM_Kiani(paramDf, n = 10000, time_scaled = FALSE, gamma = FALSE,
  agg_simus = FALSE, stimulus = c(1, 2), delta = 0.01, maxrt = 15,
  seed = NULL)
}
\arguments{
\item{paramDf}{a list or dataframe with one row. Column names should match the names of
IRM and PCRM model parameter names. For different stimulus quality/mean drift rates, names should
be v1, v2, v3,.... Additionally, the confidence thresholds should be given by names with
thetaUpper1, thetaUpper2,..., thetaLower1,... or, for symmetric thresholds only by theta1, theta2,....}

\item{n}{integer. The number of samples (per condition and stimulus direction) generated.
Total number of samples is \code{n*nConditions*length(stimulus)}.}

\item{model}{character scalar. One of "IRM" or "PCRM". ("IRMt" and "PCRMt" will also be accepted. In that case,
time_scaled is set to TRUE.)}

\item{time_scaled}{logical. Whether a time_scaled transformation for the confidence measure should
be used.}

\item{gamma}{logical. If TRUE, the gamma correlation between confidence ratings, rt and accuracy is
computed.}

\item{agg_simus}{logical. Simulation is done on a trial basis with rt's outcome. If TRUE,
the simulations will be aggregated over RTs to return only the distribution of response and
confidence ratings. Default: FALSE.}

\item{stimulus}{numeric vector. Either 1, 2 or c(1, 2) (default).
Together with condition represents the experimental situation. In a 2AFC task the presented
stimulus belongs to one of two categories. In the default setting trials with
both categories presented are simulated but one can choose to simulate only trials with the
stimulus coming from one category (each associated with positive drift in one of two accumulators).}

\item{delta}{numerical. Size of steps for the discretized simulation (see details).}

\item{maxrt}{numerical. Maximum reaction time to be simulated (see details). Default: 15.}

\item{seed}{numerical. Seeding for non-random data generation. (Also possible outside of the function.)}
}
\value{
Depending on gamma and agg_simus. If gamma is TRUE, returns a list with elements:
"simus" (the simulated data frame) and "gamma", which is again a list with elements
"condition", "rt" and "correct", each a tibble with two columns (see details for more
information). If gamma is FALSE, returns a data frame with columns: condition, stimulus,
response, correct, rt, conf (the continuous confidence measure) and rating (the discrete
confidence rating) or (if agg_simus=TRUE): condition, stimulus, response, correct, rating
and p (for the probability of a response and rating, given the condition and stimulus).
}
\description{
Simulates the decision responses, reaction times and state of the loosing accumulator
together with a discrete confidence judgment  in the independent and partially anti-correlated
race model (IRM and PCRM), given specific parameter constellations. See \code{\link{dIRM}} and
\code{\link{dPCRM}} for more information about parameters. Also computes the Gamma rank
correlation between the confidence ratings and condition (task difficulty), reaction times
and accuracy in the simulated output.
}
\details{
The simulation is done by simulating normal variables in discretized steps until
one process reaches the boundary. If no boundary is met within the maximum time, response is
set to 0. The output of the fitting function \code{\link{fitRTConf}} with the respective model
fits the argument paramDf for simulation. The Gamma coefficients are computed seperately for
correct/incorrect responses for the correlation of confidence ratings with condition and rt
and seperately for conditions for the correlation of accuracy and confidence. The resulting
tibbles in the output thus have two columns. One for the grouping variable and one for the
Gamma coefficient.
}
\note{
Different parameters for different conditions are only allowed for drift rate, \code{v},
and variability, \code{s}. All other parameters are used for all conditions.
}
\references{
Moreno-Bote, R. (2010). Decision confidence and uncertainty in diffusion models with
partially correlated neuronal integrators. Neural Computation, 22(7), 1786–1811.
https://doi.org/10.1162/neco.2010.12-08-930

Rausch, M., Hellmann, S., & Zehetleitner, M. (2018). Confidence in masked orientation judgments is
informed by both evidence and visibility. \emph{Attention, Perception, & Psychophysics}, 80(1), 134–154.
doi: 10.3758/s13414-017-1431-5
}
\author{
Sebastian Hellmann.
}
