% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getquantiles.R
\name{RTDensityToQuantiles}
\alias{RTDensityToQuantiles}
\alias{getquantiles}
\alias{getRTquantiles}
\alias{compute_quantiles_cdf}
\title{Get Quantiles from predicted Density or CDF}
\usage{
RTDensityToQuantiles(pred, p = c(0.1, 0.3, 0.5, 0.7, 0.9), agg_over = NULL,
  drop = NULL)

compute_quantiles_cdf(cdf, rt, p)
}
\arguments{
\item{pred}{dataframe. Should have at least two columns: rt (for reaction times)
and dens or densscaled. All other columns will be used as grouping factors.
As input for pred, the outputs of predictRT and predictRTModels can be used.}

\item{p}{numeric vector. Probabilities for returned quantiles. Default:
c(.1, .3, .5, .7, .9).}

\item{agg_over}{character. Names of columns to aggregate over (using the mean of
densities, which is valid only, if groups occur with equal probabilities) before
computing the quantiles.}

\item{drop}{character. Names of columns to be dropd and not used as grouping
factor or anything else in the analysis (i.e. "densscaled", if one want to enforce
the usage of an available column "dens")}

\item{cdf}{numeric. A increasing vector of the same length as \code{rt} giving the CDF for RT-Values.}

\item{rt}{numeric. A increasing vector of same length as \code{cdf}. May be anything, not only rt's.}
}
\value{
\code{RTDensityToQuantiles} returns a tibble with columns p and q indicating
probabilities and respective quantiles. Furhtermore, the output has grouping columns
identical to the additional columns in the input (without rt, dens and densscaled),
but without the ones in the agg_over argument. \code{compute_quantiles_cdf}
returns only a data.frame with columns p and q.
}
\description{
\code{compute_quantiles_cdf} computes quantiles for a given CDF vector.
\code{RTDensityToQuantiles} computes the quantiles from a vector of probability
density values within groups of other given variables (in the context of confidence
models: conditions, correct/incorrect answers and confidence ratings), if
available.
}
\details{
The RTs in the input data frame should be equidistant, i.e. the difference
between consecutive RT-values should be constant. For a reasonable accuracy the number
of steps in the input should be very high, i.e. the distant between rt values small.
The precision of the output is bounded from above by the step size of the RTs in the input.

If available, the column densscaled will be preferred (if not given in \code{drop}).
If the column dens is used, it will be scaled by the summed probability.

Attention should be given to the columns of pred other then rt and dens/densscaled, because
the quantiles are computed
}
\references{
Pleskac, T. J., & Busemeyer, J. R. (2010). Two-Stage Dynamic Signal Detection:
A Theory of Choice, Decision Time, and Confidence, \emph{Psychological Review}, 117(3),
864-901. doi:10.1037/a0019737

Rausch, M., Hellmann, S., & Zehetleitner, M. (2018). Confidence in masked orientation
judgments is informed by both evidence and visibility. \emph{Attention, Perception, &
Psychophysics}, 80(1), 134–154.  doi: 10.3758/s13414-017-1431-5
}
\author{
Sebastian Hellmann.
}
