% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWEV.R
\name{simulateWEV}
\alias{simulateWEV}
\alias{rWEV}
\alias{r2DSD}
\alias{simulate2DSD}
\title{Simulation of confidence ratings and RTs in dWEV and 2DSD confidence models}
\usage{
rWEV(paramDf, n = 10000, model = "dynWEV", simult_conf = TRUE,
  gamma = FALSE, agg_simus = FALSE, stimulus = c(-1, 1),
  method = "Cpp", precision = 3, delta = 0.01, maxrt = 15,
  seed = NULL)
}
\arguments{
\item{paramDf}{a list or dataframe with one row. Column names should match the names of dWEV and 2DSD model specific parameter names. For different stimulus quality/mean drift rates, names should be v1, v2, v3,.... Different sv parameters are possible with sv1, sv2, sv3... with equally many steps as for drift rates. Additionally, the confidence thresholds should be given by names with thetaUpper1, thetaUpper2,..., thetaLower1,... or, for symmetric thresholds only by theta1, theta2,....}

\item{n}{integer. The number of samples (per condition and stimulus direction) generated. Total number of samples is \code{n*nConditions*length(stimulus)}.}

\item{model}{character scalar. One of "WEVmu", "dynWEV", or "2DSD".}

\item{simult_conf}{logical. TRUE, if in the experiment confidence was reported simultaneously
with the decision, as then decision and confidence judgment are assumed to have happened
subsequent before response and tau is added to the simulated decision time. If FALSE,
returned response time will only be decision time plus non-judgment time component.}

\item{gamma}{logical.}

\item{agg_simus}{logical. Simulation is done on a trial basis with rt's outcome. If TRUE, the simulations will be aggregated over RTs to return only the distribution of response and confidence ratings. Default: FALSE.}

\item{stimulus}{numeric vector. Either 1, -1 or c(-1, 1) (default). Together with condition represents the experimental situation. In a 2AFC task the presented stimulus belongs to one of two categories. In the default setting trials with
both categories presented are simulated but one can choose to simulate only trials with the stimulus coming from one category (1 for the category that is associated with positive drift in the decision process where "upper"/1 responses are considered correct and -1 correspondingly for negative drifts and "lower"/-1 correct decisions).}

\item{method}{character. Method for the simulation. If "rtdists", simulations are done using the function rdiffusion from the rtdists package. Else (default), for each observation the
stochastic process is simulated in discrete steps.}

\item{precision}{\code{numerical} scalar value. Precision of calculation. Corresponds roughly to the number of decimals of the predicted CDFs that are calculated accurately. Default is 3. This argument is given directly to \code{\link[rtdists:Diffusion]{rdiffusion}} used for generating samples in the decision process.}

\item{delta}{numeric. Discretization steps for simulations with the stochastic process (used, if method!="rtdists")}

\item{maxrt}{numeric. Maximum reaction time returned, if method!="rtdists". If the simulation of the stochastic process exceeds a rt of maxrt,
the response will be set to 0 and the maxrt will be returned as rt.}

\item{seed}{numerical. Seeding for non-random data generation.}
}
\value{
Depending on gamma and agg_simus. If gamma is TRUE, returns a list with elements:
"simus" (the simulated data frame) and "gamma", which is again a list with elements
"condition", "rt" and "correct", each a tibble with two columns (see details for more
information). If gamma is FALSE, returns a data frame with columns: condition, stimulus,
response, correct, rt, conf (the continuous confidence measure) and rating (the discrete
confidence rating) or (if agg_simus=TRUE): condition, stimulus, response, correct, rating
and p (for the probability of a response and rating, given the condition and stimulus).
}
\description{
Simulates the decision responses and reaction times together with a discrete confidence judgement  in the 2DSD Model (Pleskac & Busemeyer, 2010) and the dWEV model (Hellmann & Rausch), given specific parameter constellations.
See \code{\link{dWEV}} and \code{\link{d2DSD}} for more information about parameters. Also computes the Gamma rank correlation between the confidence ratings and condition (task difficulty), reaction times and accuracy in the simulated output.
}
\details{
The function combines the random generator \code{\link[rtdists:Diffusion]{rdiffusion}} from the package \code{rtdists} for the decision process outputs and \code{rnorm} to produce the confidence measure in the respective model.
The confidence outputs are then binned according to the given thresholds. The output of the fitting function \code{\link{fitRTConf}} with the respective model fits the argument paramDf for simulation.
The Gamma coefficients are computed seperately for correct/incorrect responses for the correlation of confidence ratings with condition and rt and seperately for conditions for the correlation of accuracy and confidence. The
resulting tibbles in the output thus have two columns. One for the grouping variable and one for the Gamma coefficient.
}
\note{
Different parameters for different conditions are only allowed for drift rate, \code{v}, and drift rate variability, \code{sv}. All other parameters are used for all conditions.
}
\references{
Pleskac, T. J., & Busemeyer, J. R. (2010). Two-Stage Dynamic Signal Detection: A Theory of Choice, Decision Time, and Confidence, \emph{Psychological Review}, 117(3), 864-901. doi:10.1037/a0019737

Rausch, M., Hellmann, S., & Zehetleitner, M. (2018). Confidence in masked orientation judgments is informed by both evidence and visibility. \emph{Attention, Perception, & Psychophysics}, 80(1), 134–154.  doi: 10.3758/s13414-017-1431-5
}
\author{
Sebastian Hellmann.
}
