% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dynaViTE_formula.R
\name{simulateWEV_formula}
\alias{simulateWEV_formula}
\alias{simulate_dynaViTE_formula}
\alias{transform_const_pars_to_scale}
\title{Simulation of confidence ratings and RTs in dynaViTE with various manipulations}
\usage{
simulate_dynaViTE_formula(beta, pred_matrix = NULL, fixed = NULL,
  model_matrix = NULL, n = NULL, maxt0 = NULL, restr_tau = NULL,
  delta = 0.01, maxrt = 15, simult_conf = FALSE,
  process_results = FALSE, seed = NULL)

transform_const_pars_to_scale(beta, fixed = list(), maxt0 = NULL,
  restr_tau = Inf, simult_conf = FALSE)
}
\arguments{
\item{beta}{named vector with fitted parameters. Parameters manipulated experimentally
should have a parameter for each predictor named in the form "\emph{parameter}_\emph{predictor}".
The first argument can also be provided as a list with elements "beta", "fixed",
"model_matrix", "maxt0", and "restr_tau" providing the values for the respective arguments.}

\item{pred_matrix}{matrix providing the values for the predictors used for simulation. If NULL (default),
the arguments model_matrix and n have to be specified for then n rows (with replacement) are drawn from
the model_matrix are drawn for prediction.}

\item{fixed}{list or named vector providing the parameters, which were not fitted but fixed in the model fitting.}

\item{model_matrix}{matrix providing the values of the predictors used for model fitting (only necessary, if pred_matrix is NULL).}

\item{n}{integer. The number of samples drawn from the model_matrix (only necessary, if pred_matrix is NULL).}

\item{maxt0}{numeric. Maximum possible value for t0 in the fitting procedure. t0 represents minimal non-decision time,
so it is bounded from above by the minimal observed response time in the data. This is used to scale the fitted parameter for t0 properly.}

\item{restr_tau}{numeric or Inf. Maximum possible value for tau in the fitting procedure. tau represents post-decisional accumulation time,
and my be bounded from above in the fitting (e.g. based on observed confidence response times).
This is used to scale the fitted parameter for t0 properly.}

\item{delta}{numeric. Discretization steps for simulations with the stochastic process.}

\item{maxrt}{numeric. Maximum reaction time returned.
If the simulation of the stochastic process exceeds a rt of \code{maxrt},
the response will be set to 0 and \code{maxrt} will be returned as rt.}

\item{simult_conf}{logical. \code{TRUE}, if in the experiment confidence was reported simultaneously
with the decision, as then decision and confidence judgment are assumed to have happened
subsequent before response and tau is added to the simulated decision time. If \code{FALSE}
returned response time will only be decision time plus non-judgment time component.}

\item{process_results}{logical. Whether the output simulations should contain the final
state of the decision (and visibility) process as additional column. Default is FALSE, meaning that
no additional columns for the final process states are returned.}

\item{seed}{numerical. Seeding for non-random data generation.}
}
\value{
Returns a \code{matrix} with columns for each predictor in the model_matrix/pred_matrix and
simulated outcome variables: \code{response}, \code{rt}, \code{conf} (the continuous confidence
measure) and \code{rating} (the discrete confidence rating), and \code{dec} and \code{vis}
(only if \code{process_results=TRUE}) for the final states of accumulators in the
simulation.
}
\description{
Simulates the decision responses and reaction times together with a
discrete confidence judgment in the dynaViTE model (the 2DSD model (Pleskac & Busemeyer, 2010)
and the dynWEV model (Hellmann et al., 2023) are special cases), given
previously fitted parameters together with manipulations.
For details about the parameters, see \code{\link{dWEV}}.
}
\details{
Simulation of responses and decision times is done by simulating
normal variables in discretized steps until the lower or upper boundary
is met (or the maximal rt is reached). Afterwards, a confidence measure
is simulated according to the respective model.

The confidence outputs are then binned according to the given thresholds.
The output of the fitting function \code{\link{fitRTConf_formula}} fits the argument \code{beta} for simulation.
}
\note{
Different parameters for different conditions are only allowed for drift rate,
\code{v}, drift rate variability, \code{sv} and diffusion constant \code{s}.
All other parameters are used for all conditions.
}
\examples{
# Examples for "dynWEV" model (equivalent applicable
# for "2DSD" model (with less parameters))
# 1. Define some parameter set in a data.frame
paramDf <- data.frame(a=2.5,v1=0.1, v2=1, t0=0.1,z=0.55,
                      sz=0.3,sv=0.8, st0=0,  tau=3, w=0.1,
                      theta1=0.8, svis=0.5, sigvis=0.8)

# 2. Simulate trials for both stimulus categories and all conditions (2)
simus <- simulateWEV(paramDf, model="dynWEV")
head(simus)
\donttest{
  library(ggplot2)
  simus <- simus[simus$response!=0,]
  simus$rating <- factor(simus$rating, labels=c("unsure", "sure"))
  ggplot(simus, aes(x=rt, group=interaction(correct, rating),
                    color=as.factor(correct), linetype=rating))+
    geom_density(linewidth=1.2)+xlim(c(0,5))+
    facet_grid(rows=vars(condition), labeller = "label_both")
}

# automatically aggregate simulation distribution
# to get only accuracy x confidence rating distribution for
# all conditions
agg_simus <- simulateWEV(paramDf, model="dynWEV", agg_simus = TRUE)
head(agg_simus)
\donttest{
  agg_simus$rating <- factor(agg_simus$rating, labels=c("unsure", "sure"))
  library(ggplot2)
  ggplot(agg_simus, aes(x=rating, group=correct, fill=as.factor(correct), y=p))+
    geom_bar(stat="identity", position="dodge")+
    facet_grid(cols=vars(condition), labeller = "label_both")
}
\donttest{
  # Compute Gamma correlation coefficients between
  # confidence and other behavioral measures
  # output will be a list
  simu_list <- simulateWEV(paramDf,n = 400, model="dynWEV", gamma=TRUE)
  simu_list
}
}
\references{
Hellmann, S., Zehetleitner, M., & Rausch, M. (2023). Simultaneous modeling of choice, confidence and response time in visual perception. \emph{Psychological Review} 2023 Mar 13. doi: 10.1037/rev0000411. Epub ahead of print. PMID: 36913292.
}
\author{
Sebastian Hellmann.
}
