% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynConfiR-package.R
\docType{package}
\name{dynConfiR-package}
\alias{dynConfiR-package}
\alias{dynWEV-package}
\title{The dynConfiR Package}
\description{
Dynamic Models for Confidence and Response Time Distributions
}
\details{
\tabular{ll}{
Package: \tab dynConfiR\cr
Type: \tab Package\cr
Version: \tab 0.0.2\cr
Date: \tab 2022-01-12\cr
Depends: \tab R (>= 4.0)\cr
License: \tab GPL (>=3)\cr
URL: \tab https://github.com/SeHellmann/dynConfiR\cr
}

Provides response time and confidence distributions (density/PDF) for following models: 2DSD, dynWEV, IRM and PCRM
}
\author{
Sebastian Hellmann
}
\keyword{package}
