% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictratingdist_RM.R
\name{predictRM}
\alias{predictRM}
\alias{predictRM_Conf}
\alias{predictIRM}
\alias{predictPCRM}
\alias{predictRM_RT}
\title{Prediction of Rating and Reaction Time Distribution in  the independent
and partially anti-correlated race models of confidence}
\usage{
predictRM_Conf(paramDf, model = "IRM", time_scaled = FALSE, maxrt = 15,
  subdivisions = 100L, stop.on.error = FALSE, .progress = TRUE)

predictRM_RT(paramDf, model = "IRM", time_scaled = FALSE, maxrt = 9,
  subdivisions = 100L, minrt = NULL, scaled = FALSE, DistConf = NULL,
  .progress = TRUE)
}
\arguments{
\item{paramDf}{a list or dataframe with one row. Column names should match the
names of dRM model parameter names. For different stimulus
quality/mean drift rates, names should be v1, v2, v3,.... Different s parameters
are possible with s1, s2, s3... with equally many steps as for drift rates.
Additionally, the confidence thresholds should be given by names with
thetaUpper1, thetaUpper2,..., thetaLower1,... or,
for symmetric thresholds only by theta1, theta2,....}

\item{model}{character scalar. One of "IRM" or "PCRM". ("IRMt" and "PCRMt" will also be accepted. In that case,
time_scaled is set to TRUE.)}

\item{time_scaled}{logical. Whether the confidence measure should be scaled by 1/sqrt(rt). Default: FALSE.
(It is set to TRUE, if model is "IRMt" or "PCRMt")}

\item{maxrt}{numeric. The maximum RT for the
integration/density computation. Default: 15 (for _Conf (integration)), 9 (for _RT).}

\item{subdivisions}{\code{integer} (default: 100).
For \code{predictRM_Conf} it is used as argument for the inner integral routine.
For \code{predictRM_RT} it is the number of points for which the density is computed.}

\item{stop.on.error}{logical. Argument directly passed on to integrate. Default is FALSE, since
the densities invoked may converge slowly (but are still quite accurate) which causes R to throw an
error.}

\item{.progress}{logical. If TRUE (default) a progress bar is drawn to the console.}

\item{minrt}{numeric or NULL(default). The minimum rt for the density computation.}

\item{scaled}{logical. For \code{predictRM_RT}. Whether the computed density
should be scaled to integrate to one (additional column densscaled). Otherwise the output
is a defective density (i.e. its integral is equal to the probability of a response and
not 1). If TRUE, the argument DistConf should be given, if available. Default: FALSE.}

\item{DistConf}{NULL or data.frame. For \code{predictRM_RT}. A data.frame or matrix
with column names, giving the distribution of response and rating choices for
different conditions and stimulus categories in the form of the output of
\code{predictRM_Conf}. It is only necessary, if scaled=TRUE, because these
probabilities are used for scaling. If scaled=TRUE and DistConf=NULL, it will be computed
with the function \code{predictRM_Conf}, which takes some time and the function will
throw a message. Default: NULL}
}
\value{
\code{predictRM_Conf} gives a data frame/tibble with columns: condition, stimulus,
response, rating, correct, p, info, err. p is the predicted probability of a response
and rating, given the stimulus category and condition. Message and error refer to the
respective outputs of the integration routine used for prediction.
\code{predictRM_RT} returns a data frame/tibble with columns: condition, stimulus,
response, rating, correct, rt and dens (and densscaled, if scaled=TRUE).
}
\description{
\code{predictRM_Conf} predicts the categorical response distribution of
decision and confidence ratings, \code{predictRM_RT} computes the predicted
RT-Distribution (density) in the independent and partially anti-correlated race models
(IRM and PCRM), given specific parameter constellations.
See \code{\link{dIRM}} and \code{\link{dPCRM}} for more information about parameters.
}
\details{
The function \code{predictRM_Conf} consists merely of an integration of
the reaction time density, \code{\link{dIRM}} and \code{\link{dPCRM}}, over the reaction time in a reasonable
interval (0 to maxrt). The function \code{predictRM_RT} wraps these density
functions to a parameter set input and a data.frame output.
For the argument \code{paramDf}, the output of the fitting function \code{\link{fitRTConf}}
with the respective model may be used.
}
\note{
Different parameters for different conditions are only allowed for drift rate,
\code{v}, and drift rate variability, \code{s}. All other parameters are used for all
conditions.
}
\references{
Moreno-Bote, R. (2010). Decision confidence and uncertainty in diffusion models with
partially correlated neuronal integrators. Neural Computation, 22(7), 1786–1811.
https://doi.org/10.1162/neco.2010.12-08-930
}
\author{
Sebastian Hellmann.
}
